@@ -1,3 +1,3 @@
 module gonetlib
 
-go 1.17
+go 1.18
